requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
contourpy==1.3.3
cycler==0.12.1
distorch @ git+https://github.com/jeromerony/distorch.git@80f190cb46632b1f05dc1a608f6b368f89dbcea2
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
nibabel==5.3.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-ranger==0.1.1
pytz==2025.2
PyYAML==6.0.3
scikit-image==0.25.2
scipy==1.16.2
simsimd==6.5.3
six==1.17.0
stringzilla==4.0.14
sympy==1.14.0
tifffile==2025.9.20
torch==2.8.0
torch-optimizer==0.3.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
